+++ tt_metal/impl/dispatch/command_queue.hpp
@@ -331,16 +331,6 @@ class EnqueueProgramCommand : public Command {
         uint32_t unicast_cores_launch_message_wptr,
         SubDeviceId sub_device_id);
 
-    void assemble_preamble_commands(
-        ProgramCommandSequence& program_command_sequence, const tt::stl::Span<ConfigBufferEntry> kernel_config_addrs);
-    void assemble_stall_commands(ProgramCommandSequence& program_command_sequence, bool prefetch_stall);
-    void assemble_runtime_args_commands(ProgramCommandSequence& program_command_sequence);
-    void assemble_device_commands(
-        ProgramCommandSequence& program_command_sequence, const tt::stl::Span<ConfigBufferEntry> kernel_config_addrs);
-    void update_device_commands(
-        ProgramCommandSequence& cached_program_command_sequence,
-        const tt::stl::Span<ConfigBufferEntry> kernel_config_addrs);
-
     void write_program_command_sequence(
         const ProgramCommandSequence& program_command_sequence, bool stall_first, bool stall_before_program);
 
@@ -527,7 +517,6 @@ class HWCommandQueue {
     void set_num_worker_sems_on_dispatch(uint32_t num_worker_sems);
     void set_go_signal_noc_data_on_dispatch(const vector_memcpy_aligned<uint32_t>& go_signal_noc_data);
     void reset_worker_state(bool reset_launch_msg_state);
-
    private:
     uint32_t id;
     uint32_t size_B;
