+++ tt_metal/impl/dispatch/kernel_config/dispatch_s.cpp
@@ -27,7 +27,7 @@ void DispatchSKernel::GenerateStaticConfigs() {
             dispatch_s_buffer_base = dispatch_buffer_base;
         }
     }
-    logical_core_ = dispatch_core_manager::instance().dispatcher_s_core(device_->id(), channel, cq_id_);
+
     static_config_.cb_base = dispatch_s_buffer_base;
     static_config_.cb_log_page_size = dispatch_constants::DISPATCH_S_BUFFER_LOG_PAGE_SIZE;
     static_config_.cb_size = my_dispatch_constants.dispatch_s_buffer_size();
